moving app file
