+++ db/connection.js
@@ -12,6 +12,8 @@ if( ENV === "production"){
   config.connectionString = process.env.DATABASE_URL;
   config.max= 2;
 };
+
+console.log(config)
 if (!process.env.PGDATABASE && !process.env.DATABASE_URL) {
   throw new Error('PGDATABASE not set or DATABASE_URL not set');
 };
