+++ listen.js
@@ -1,4 +1,4 @@
-const app = require('./MVC/app')
+const app = require('./app')
 const { PORT = 9090 } = process.env;
 
 app.listen(PORT, () => console.log(`Listening on ${PORT}...`));
